+++ lib/rack/dedos/filters/spamhaus.rb
@@ -1,0 +1,40 @@
+# frozen_string_literal: true
+
+require 'resolv'
+
+module Rack
+  module Dedos
+    module Filters
+      class Spamhaus < Base
+
+        QUERY_DOMAIN = 'zen.spamhaus.org'.freeze
+
+        def name
+          :spamhaus
+        end
+
+        def allowed?(request, ip)
+          resolver.getresources(domain_for(ip), Resolv::DNS::Resource::IN::A).empty?
+        rescue => error
+          logger.error("request from #{ip} allowed due to error: #{error.message}")
+          true
+        end
+
+        private
+
+        def resolver
+          Thread.current.thread_variable_get(:spamhaus_resolver) ||
+            Thread.current.thread_variable_set(
+              :spamhaus_resolver,
+              Resolv::DNS.new.tap { _1.timeouts = [1, 2] }
+            )
+        end
+
+        def domain_for(ip)
+          ip.split('.').reverse.join('.').concat('.', QUERY_DOMAIN)
+        end
+
+      end
+    end
+  end
+end